requirements.txt
anyio==4.2.0
asgiref==3.7.2
asyncio==3.4.3
certifi==2024.2.2
Django==4.0.4
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.2.0
python-dotenv==1.0.1
python-telegram-bot==20.8
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
